Created a Powershell version of gen-djvu-dsed
This version has all the capabilities of the perl6 version, so
it is interesting to compare them for clarity.  It's obvious
that the perl is cleaner-looking,  Powershell has a lot of
noise for things like scoping and parsing ints that perl just
does automatically.  On the other hand, the Powershell version
is a lot faster (due to startup-time, mostly, I'm sure).